Parser architecture + all requirements
