+++ Markdown/MarkdownTests/MdTests.cs
@@ -14,35 +14,35 @@ namespace MarkdownTests
         [SetUp]
         public void SetUp()
         {
-            parser = new Md();
+            parser = new Md(new MarkdownParser());
         }
 
-        [TestCase("aa_bb_aa", @"aa<em>bb</em>aa")]
+        [TestCase("aa _bb_ aa", @"aa <em>bb</em> aa")]
         [TestCase("_abc_", @"<em>abc</em>")]
         [TestCase(" _abc_", @" <em>abc</em>")]
         [TestCase("_abc_ ", @"<em>abc</em> ")]
         [TestCase("_abc_ d _abc_", @"<em>abc</em> d <em>abc</em>")]
         public void ShouldParse_Italic(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
 
         [TestCase(@"\_asd\_", "_asd_")]
         [TestCase(@" \_ ", " _ ")]
-        [TestCase(@" \_\__a_ \_", @" __<em>a</em> _")]
+        [TestCase(@" \_ _\_a_ \_", @" _ <em>_a</em> _")]
         [TestCase(@" \_abcdefg\_ ", " _abcdefg_ ")]
         //[TestCase(@"\", @"\")]
         public void ShouldParse_EscapeSymbols(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
 
         [TestCase("__abc__", @"<strong>abc</strong>")]
         public void ShouldParse_Strong(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
 
@@ -50,7 +50,7 @@ namespace MarkdownTests
         [TestCase(@"__abc__ _abc_", @"<strong>abc</strong> <em>abc</em>")]
         public void ShouldParse_StrongAndItalic(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
         
@@ -58,7 +58,7 @@ namespace MarkdownTests
         [TestCase(@"__abc _cde_ _cde_ abc__", @"<strong>abc <em>cde</em> <em>cde</em> abc</strong>")]
         public void ShouldParse_ItalicInStrong(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
 
@@ -66,7 +66,7 @@ namespace MarkdownTests
         [TestCase(@"_abc __cde__ __cde__ abc_", @"<em>abc __cde__ __cde__ abc</em>")]
         public void ShouldParse_StrongInItalic_AsSymbols(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
 
@@ -77,7 +77,15 @@ namespace MarkdownTests
         [TestCase("_ __a _", "<em> __a </em>")]
         public void ShouldParse_NotClosed_AsSymbols(string rowString, string expected)
         {
-            var result = parser.Render(rowString, Markups.Markdown, Markups.Html);
+            var result = parser.Render(rowString, Markups.Html);
+            result.Should().BeEquivalentTo(expected);
+        }
+
+        [TestCase("a_a_", "a_a_")]
+        [TestCase("_a_a", "_a_a")]
+        public void ShouldParse_OnlyTagsWithWhitespacesAtTheEdges(string rowString, string expected)
+        {
+            var result = parser.Render(rowString, Markups.Html);
             result.Should().BeEquivalentTo(expected);
         }
     }
